SetupHelper#removeServerTestProfile checks whether directories exist before it tries to delete them

+++ org.eclipse.emf.emfstore.client.test/src/org/eclipse/emf/emfstore/client/test/SetupHelper.java
@@ -696,11 +696,17 @@ public class SetupHelper {
 	 */
 	public static void removeServerTestProfile() throws IOException {
 		String serverPath = ServerConfiguration.getServerHome();
-		File serverDirectory = new File(serverPath);
-		FileUtils.deleteDirectory(serverDirectory);
 		String clientPath = Configuration.getWorkspaceDirectory();
+		File serverDirectory = new File(serverPath);
 		File clientDirectory = new File(clientPath);
-		FileUtils.deleteDirectory(clientDirectory);
+
+		if (serverDirectory.exists()) {
+			FileUtils.deleteDirectory(serverDirectory);
+		}
+
+		if (clientDirectory.exists()) {
+			FileUtils.deleteDirectory(clientDirectory);
+		}
 	}
 
 }
